refactor(script/release): packaged changes are now committed by default
Previously, when the `release` script would package any changes to the action down to the `./dist/index.js` file and then ask if the user wanted to commit those changes it would default to `N` instead of `Y` for the response. This commit flips that and now `Y` is the default so I can just press enter to commit the latest version of the packaged script.