feat(release.sh): added ability to provide new tag as arg
This commit makes it so you can now run the release script and provide it an argument that'll be used as the `new_tag` variable instead of prompting you for it. For example, you can now do the following:

`./script/release v0.0.0` and it'll release a new version with the tag `v0.0.0`.